mds: do not allow GLAZYIO in mix->sync state
GLAZYIO is not allowed in SYNC, so we cannot allow it in the preceding
gather state.

I verified the other GLAZYIO rules look ok.  We should make a validater
to confirm that no gather state includes caps that its target state
does not... or at least assert as much in eval_gather().

Backport: cuttlefish
Signed-off-by: Sage Weil <sage@inktank.com>